feat: Support integer formatting notations
Add `Formatted<i64>::format()`, which takes an `IntegerFormat` and
formats the integer accordingly. This method returns false if the format
cannot be applied to the current value (e.g. binary notation for a
negative numbers).

The `IntegerFormat` struct can be customized with `notation()`, which
takes an `IntegerFormatNotation` enum value. Both the `IntegerFormat`
struct and `IntegerFormatNotation` enum are marked as `non_exhaustive`,
so that new notations (e.g. hexadecimal with mixed case) may be added in
the future without breaking compatibility.